pyproject.toml
[project]
name = "hl-vox-timeleft"
version = "0.1.0"
description = "Half-Life Pomodoro Timer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dearpygui>=2.0.0",
    "num2words>=0.5.14",
    "pygame>=2.5.0",
    "google-cloud-bigquery>=3.0.0",
]
